backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-gis>=1.0,<2.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# OpenAPI Documentation
drf-spectacular>=0.27,<1.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Environment
python-dotenv>=1.0,<2.0

# Redis & Caching
django-redis>=5.4,<6.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code Quality
ruff>=0.3,<1.0
black>=24.0,<25.0
pre-commit>=3.6,<4.0

# Utilities
Pillow>=10.0,<11.0
python-dateutil>=2.8,<3.0
